Add disable flag to ssh tests
* Refactor environment flags
* Add skip flags for ssh tests

+++ cmd/kick/main.go
@@ -31,14 +31,17 @@ func main() {
 	inject := di.Setup(home)
 
 	// open log file and close on exit
-	lf := inject.MakeLogFile()
-	defer func() {
-		lf.Close()
-		info, err := os.Stat(lf.Name())
-		if err != nil && info.Size() == 0 {
-			os.Remove(lf.Name())
-		}
-	}()
+	logfile := os.Getenv("KICK_LOG")
+	if logfile != "" {
+		lf := inject.MakeLogFile(logfile)
+		defer func() {
+			lf.Close()
+			info, err := os.Stat(lf.Name())
+			if err != nil && info.Size() == 0 {
+				os.Remove(lf.Name())
+			}
+		}()
+	}
 
 	if os.Getenv("KICK_DEBUG") == "true" {
 		inject.LogLevel(logger.DebugLevel)
